move systems design with system state
