+++ modules/core/src/ship/ship-state.ts
@@ -8,7 +8,6 @@ import { ModelParams } from '../model-params';
 import { Radar } from './radar';
 import { Reactor } from './reactor';
 import { ShipDirection } from './ship-direction';
-import { ShipPropertiesDesign } from './ship-configuration';
 import { SmartPilot } from './smart-pilot';
 import { Spaceship } from '../space';
 import { Thruster } from './thruster';
@@ -18,6 +17,13 @@ export enum TargetedStatus {
     LOCKED,
     FIRED_UPON,
 }
+
+export type ShipPropertiesDesign = {
+    rotationCapacity: number;
+    rotationEnergyCost: number;
+    maxChainGunAmmo: number;
+};
+
 @rangeSchema({ '/turnSpeed': [-90, 90], '/angle': [0, 360] })
 export class ShipState extends Spaceship {
     @type(ModelParams)
